pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hidemyhtml"
version = "1.0.0"
description = "Obfuscates an HTML page to avoid pesky scanners"
authors = [
  { name = "Fudgedotdotdot" },
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
  "calmjs.parse==1.3.1",
  "colorama==0.4.6"
]

[tool.setuptools.packages.find]
include = ["hidemyhtml"]

[tool.setuptools.package-data]
hidemyhtml = ["template/**"]

[project.scripts]
hidemyhtml = "hidemyhtml.hidemyhtml:main"
